// reportus-frontend/src/payload-types.ts
export interface Customer {
  id: string;
  name: string;
  contact?:
    | {
        name?: string | null;
        phone?: string | null;
        email?: string | null;
        role?: string | null;
        id?: string | null;
      }[]
    | null;
  notes?: string | null;
  _status: "active" | "archived";
  updatedAt: string;
  createdAt: string;
}
// ...
export interface ExternalAlert {
  id: string;
  customer: string | Customer;
  reference_id: string;
  asset: {
    host: string;
    alias?: string | null;
  };
  level: "MEDIUM" | "HIGH" | "CRITICAL" | "HIGH CRITICAL";
  description?: string | null;
  notes?: string | null;
  updatedAt: string;
  createdAt: string;
}